do not try rescaning paths at the internal app Files Directory (recycle bin)

+++ commons/src/main/kotlin/com/simplemobiletools/commons/extensions/Context-storage.kt
@@ -375,6 +375,11 @@ fun Context.getOTGItems(path: String, shouldShowHidden: Boolean, getProperFileSi
 }
 
 fun Context.rescanDeletedPath(path: String, callback: (() -> Unit)? = null) {
+    if (path.startsWith(filesDir.toString())) {
+        callback?.invoke()
+        return
+    }
+
     if (deleteFromMediaStore(path)) {
         callback?.invoke()
     } else {
@@ -383,13 +388,13 @@ fun Context.rescanDeletedPath(path: String, callback: (() -> Unit)? = null) {
             return
         }
 
-        MediaScannerConnection.scanFile(applicationContext, arrayOf(path), null, { s, uri ->
+        MediaScannerConnection.scanFile(applicationContext, arrayOf(path), null) { s, uri ->
             try {
                 applicationContext.contentResolver.delete(uri, null, null)
             } catch (e: Exception) {
             }
             callback?.invoke()
-        })
+        }
     }
 }
